# Avoid unnecessary retry delay in cluster client following MOVED and ASK redirection